feat(spa): split persistent/static entrypoints
