+++ packages/integrations/spa/client-static.js
@@ -1,0 +1,8 @@
+import listen from 'micromorph/spa';
+
+export default () =>
+	listen({
+		afterDiff() {
+			window.dispatchEvent(new CustomEvent('astro:locationchange'));
+		},
+	});
